third_party/blink/web_tests/fast/css-grid-layout/grid-auto-repeat-huge-grid-008.html

<!DOCTYPE html>
<title>Test for auto-fit and auto-fill with huge grids (lots of tracks)</title>
<link href="resources/grid.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/height-keyword-classes.css" rel="stylesheet">
<link href="resources/huge-grids.css" rel="stylesheet">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/grid-tracks-length.js"></script>

<div id="tallAutoFillGridReversed" class="grid tallGrid lotsOfFixedRepeatWithAutoFillRowsReversed"></div>

<script>

test(function() {
     var autoFillGrid = testElement("tallAutoFillGridReversed", "grid-template-rows", 3570505);

     assert_equals(autoFillGrid[3570505 - 1], "20px");
     assert_equals(autoFillGrid[3570505 - 2], "7px");
     assert_equals(autoFillGrid[0], "1px");
     assert_equals(autoFillGrid[1], "1px");

}, "Test that we don't get more than kGridMaxTracks repetitions even on very tall grids (auto repeat tracks clamped).");

</script>
